+++ CHANGES.md
@@ -1,6 +1,13 @@
 Changes
 =======
 
+0.4.1
+=====
+
+* 2021-03-19 The request-scoped services were not quite correct as they could have been instantiated twice
+  - once in the traversal-time and the other time after context was set. Now the context is forced to None
+  for the request-scoped services.
+
 0.4.0
 =====